[FP-50] Rename: 오류나는 파일 이름 변경
